AOE Spring 2022 release

+++ Deploy-AzureOptimizationEngine.ps1
@@ -9,7 +9,13 @@ param (
     [string] $ArtifactsSasToken,
 
     [Parameter(Mandatory = $false)]
-    [switch] $DoPartialUpgrade # updates only storage account containers, Automation assets and SQL Database model
+    [switch] $DoPartialUpgrade,
+
+    [Parameter(Mandatory = $false)]
+    [switch] $IgnoreNamingAvailabilityErrors,
+
+    [Parameter(Mandatory = $false)]
+    [string] $SilentDeploymentSettingsPath
 )
 
 function ConvertTo-Hashtable {
@@ -45,6 +51,8 @@ function ConvertTo-Hashtable {
 }
 
 $ErrorActionPreference = "Stop"
+
+#region Deployment environment settings
 
 $lastDeploymentStatePath = ".\last-deployment-state.json"
 $deploymentOptions = @{}
@@ -112,6 +120,10 @@ else {
     }
 }
 
+#endregion
+
+#region Azure subscription choice
+
 Write-Host "Getting Azure subscriptions (filtering out unsupported ones)..." -ForegroundColor Green
 
 $subscriptions = Get-AzSubscription | Where-Object { $_.State -eq "Enabled" -and $_.SubscriptionPolicies.QuotaId -notlike "Internal*" -and $_.SubscriptionPolicies.QuotaId -notlike "AAD*" }
@@ -174,6 +186,9 @@ if ($ctx.Subscription.Id -ne $subscriptionId) {
     $ctx = Select-AzSubscription -SubscriptionId $subscriptionId
 }
 
+#endregion
+
+#region Resource naming options
 $workspaceReuse = $null
 
 $deploymentNameTemplate = "{0}" + (Get-Date).ToString("yyMMddHHmmss")
@@ -260,7 +275,9 @@ else
     $laWorkspaceName = $deploymentOptions["WorkspaceName"]        
     $deploymentName = $deploymentNameTemplate -f $resourceGroupName
 }
+#endregion
 
+#region Resource naming availability checks
 Write-Host "Checking name prefix availability..." -ForegroundColor Green
 
 Write-Host "...for the Storage Account..." -ForegroundColor Green
@@ -315,13 +332,15 @@ else {
     Write-Host "(The SQL Server was already deployed)" -ForegroundColor Green
 }
 
-if (-not($nameAvailable))
+if (-not($nameAvailable) -and -not($IgnoreNamingAvailabilityErrors))
 {
     throw "Please, fix naming issues. Terminating execution."
 }
 
 Write-Host "Chosen resource names are available for all services" -ForegroundColor Green
+#endregion
 
+#region Additional resource options (LA reused, region, SQL user)
 if (-not($deploymentOptions["WorkspaceResourceGroupName"]))
 {
     if ("Y", "y" -contains $workspaceReuse) {
@@ -383,7 +402,9 @@ else
     $sqlAdmin = $deploymentOptions["SqlAdmin"]    
 }
 $sqlPass = Read-Host "Please, input the SQL Admin ($sqlAdmin) password" -AsSecureString
+#endregion
 
+#region Partial upgrade dependent resource checks
 if (-not($DoPartialUpgrade))
 {
     $upgrading = $false
@@ -415,8 +436,11 @@ else
         }
         else
         {
-            $upgrading = $false    
-            Write-Host "Did not find the $sqlServerName SQL Server." -ForegroundColor Yellow
+            if (-not($IgnoreNamingAvailabilityErrors))
+            {
+                $upgrading = $false    
+                Write-Host "Did not find the $sqlServerName SQL Server." -ForegroundColor Yellow    
+            }
         }
     
         $auto = Get-AzAutomationAccount -ResourceGroupName $resourceGroupName -Name $automationAccountName -ErrorAction SilentlyContinue
@@ -441,6 +465,7 @@ else
         $upgrading = $false    
     }        
 }
+#endregion
 
 $deploymentMessage = "Deploying Azure Optimization Engine to subscription"
 if ($upgrading)
@@ -454,6 +479,7 @@ if ("Y", "y" -contains $continueInput) {
 
     $deploymentOptions | ConvertTo-Json | Out-File -FilePath $lastDeploymentStatePath -Force
     
+    #region Computing schedules base time
     $baseTime = (Get-Date).ToUniversalTime().ToString("u")
     $upgradingSchedules = $false
     $schedules = Get-AzAutomationSchedule -ResourceGroupName $resourceGroupName -AutomationAccountName $automationAccountName -ErrorAction SilentlyContinue
@@ -469,9 +495,11 @@ if ("Y", "y" -contains $continueInput) {
     else {
         Write-Host "Automation schedules base time automatically set to $baseTime." -ForegroundColor Green
     }
+    #endregion
 
     if (-not($upgrading))
     {
+        #region Template-based deployment
         $jobSchedules = Get-AzAutomationScheduledRunbook -ResourceGroupName $resourceGroupName -AutomationAccountName $automationAccountName -ErrorAction SilentlyContinue
         if ($jobSchedules.Count -gt 0) {
             Write-Host "Unregistering previous runbook schedules associations from $automationAccountName..." -ForegroundColor Green
@@ -501,9 +529,11 @@ if ("Y", "y" -contains $continueInput) {
                 -sqlAdminLogin $sqlAdmin -sqlAdminPassword $sqlPass -artifactsLocationSasToken (ConvertTo-SecureString $ArtifactsSasToken -AsPlainText -Force)        
         }
         $spnId = $deployment.Outputs['automationPrincipalId'].Value 
+        #endregion
     }
     else
     {
+        #region Partial upgrade deployment
         $upgradeManifest = Get-Content -Path "./upgrade-manifest.json" | ConvertFrom-Json
         Write-Host "Creating missing storage account containers..." -ForegroundColor Green
         $upgradeContainers = $upgradeManifest.dataCollection.container
@@ -647,7 +677,7 @@ if ("Y", "y" -contains $continueInput) {
                     "Export" {
                         $hybridWorkerName = $exportHybridWorkerOption
                     }
-                    "Recommmend" {
+                    "Recommend" {
                         $hybridWorkerName = $recommendHybridWorkerOption
                     }
                     "Ingest" {
@@ -691,7 +721,7 @@ if ("Y", "y" -contains $continueInput) {
                         "Export" {
                             $hybridWorkerName = $exportHybridWorkerOption
                         }
-                        "Recommmend" {
+                        "Recommend" {
                             $hybridWorkerName = $recommendHybridWorkerOption
                         }
                         "Ingest" {
@@ -737,7 +767,7 @@ if ("Y", "y" -contains $continueInput) {
                     "Export" {
                         $hybridWorkerName = $exportHybridWorkerOption
                     }
-                    "Recommmend" {
+                    "Recommend" {
                         $hybridWorkerName = $recommendHybridWorkerOption
                     }
                     "Ingest" {
@@ -789,8 +819,10 @@ if ("Y", "y" -contains $continueInput) {
         {
             Remove-AzAutomationRunbook -AutomationAccountName $automationAccountName -Name $deprecatedRunbook -ResourceGroupName $resourceGroupName -Force -ErrorAction SilentlyContinue
         }
+        #endregion
     }
 
+    #region Schedules reset
     if ($upgradingSchedules) {
         $schedules = Get-AzAutomationSchedule -ResourceGroupName $resourceGroupName -AutomationAccountName $automationAccountName
         $dailySchedules = $schedules | Where-Object { $_.Frequency -eq "Day" -or $_.Frequency -eq "Hour" }
@@ -819,13 +851,9 @@ if ("Y", "y" -contains $continueInput) {
             Invoke-AzRestMethod -Path $automationPath -Method PUT -Payload $body | Out-Null
         }
     }
-        
-    $myPublicIp = (Invoke-WebRequest -uri "http://ifconfig.me/ip").Content
-
-    Write-Host "Opening SQL Server firewall temporarily to your public IP ($myPublicIp)..." -ForegroundColor Green
-    $tempFirewallRuleName = "InitialDeployment"            
-    New-AzSqlServerFirewallRule -ResourceGroupName $resourceGroupName -ServerName $sqlServerName -FirewallRuleName $tempFirewallRuleName -StartIpAddress $myPublicIp -EndIpAddress $myPublicIp -ErrorAction SilentlyContinue
+    #endregion
     
+    #region Deployment date Automation variable
     Write-Host "Checking Azure Automation variable referring to the initial Azure Optimization Engine deployment date..." -ForegroundColor Green
     $deploymentDateVariableName = "AzureOptimization_DeploymentDate"
     $deploymentDateVariable = Get-AzAutomationVariable -ResourceGroupName $resourceGroupName -AutomationAccountName $automationAccountName -Name $deploymentDateVariableName -ErrorAction SilentlyContinue
@@ -836,11 +864,31 @@ if ("Y", "y" -contains $continueInput) {
         New-AzAutomationVariable -Name $deploymentDateVariableName -Description "The date of the initial engine deployment" `
             -ResourceGroupName $resourceGroupName -AutomationAccountName $automationAccountName -Value $deploymentDate -Encrypted $false
     }
+    #endregion
 
+    #region Open SQL Server firewall rule
+    if (-not($sqlServerName -like "*.database.*"))
+    {
+        $myPublicIp = (Invoke-WebRequest -uri "http://ifconfig.me/ip").Content
+
+        Write-Host "Opening SQL Server firewall temporarily to your public IP ($myPublicIp)..." -ForegroundColor Green
+        $tempFirewallRuleName = "InitialDeployment"            
+        New-AzSqlServerFirewallRule -ResourceGroupName $resourceGroupName -ServerName $sqlServerName -FirewallRuleName $tempFirewallRuleName -StartIpAddress $myPublicIp -EndIpAddress $myPublicIp -ErrorAction SilentlyContinue    
+    }
+    #endregion
+    
+    #region SQL Database model deployment
     Write-Host "Deploying SQL Database model..." -ForegroundColor Green
     
-    $sqlPassPlain = (New-Object PSCredential "user", $sqlPass).GetNetworkCredential().Password        
-    $sqlServerEndpoint = "$sqlServerName$($cloudDetails.SqlDatabaseDnsSuffix)"
+    $sqlPassPlain = (New-Object PSCredential "user", $sqlPass).GetNetworkCredential().Password     
+    if (-not($sqlServerName -like "*.database.*"))
+    {
+        $sqlServerEndpoint = "$sqlServerName$($cloudDetails.SqlDatabaseDnsSuffix)"
+    }
+    else 
+    {
+        $sqlServerEndpoint = $sqlServerName
+    }
     $databaseName = $sqlDatabaseName
     $SqlTimeout = 60
     $tries = 0
@@ -950,10 +998,17 @@ if ("Y", "y" -contains $continueInput) {
     if (-not($connectionSuccess)) {
         throw "Could not establish connection to SQL."
     }
+    #endregion
     
-    Write-Host "Deleting temporary SQL Server firewall rule..." -ForegroundColor Green
-    Remove-AzSqlServerFirewallRule -FirewallRuleName $tempFirewallRuleName -ResourceGroupName $resourceGroupName -ServerName $sqlServerName    
+    #region Close SQL Server firewall rule
+    if (-not($sqlServerName -like "*.database.*"))
+    {
+        Write-Host "Deleting temporary SQL Server firewall rule..." -ForegroundColor Green
+        Remove-AzSqlServerFirewallRule -FirewallRuleName $tempFirewallRuleName -ResourceGroupName $resourceGroupName -ServerName $sqlServerName        
+    }    
+    #endregion
 
+    #region Workbooks deployment
     Write-Host "Publishing workbooks..." -ForegroundColor Green
     $workbooks = Get-ChildItem -Path "./views/workbooks/" | Where-Object { $_.Name.EndsWith("-arm.json") }
     $la = Get-AzOperationalInsightsWorkspace -ResourceGroupName $laWorkspaceResourceGroup -Name $laWorkspaceName
@@ -969,7 +1024,9 @@ if ("Y", "y" -contains $continueInput) {
             Write-Host "Failed to deploy the workbook. If you are upgrading AOE, please remove first the $($armTemplate.parameters.workbookDisplayName.defaultValue) workbook from the $laWorkspaceName Log Analytics workspace and then re-deploy." -ForegroundColor Yellow            
         }
     }
-    
+    #endregion
+
+    #region Grant Azure AD role to AOE principal
     if ($null -eq $spnId)
     {
         $auto = Get-AzAutomationAccount -Name $automationAccountName -ResourceGroupName $resourceGroupName
@@ -1027,11 +1084,22 @@ if ("Y", "y" -contains $continueInput) {
         if (-not($globalReaders -contains $spnId))
         {
             New-MgDirectoryRoleMemberByRef -DirectoryRoleId $globalReaderRole.Id -BodyParameter @{"@odata.id" = "https://graph.microsoft.com/v1.0/directoryObjects/$spnId"}
-            Write-Host "Role granted." -ForegroundColor Green
+            Start-Sleep -Seconds 5
+            $globalReaderRole = Get-MgDirectoryRole -ExpandProperty Members -Property Id,Members,DisplayName,RoleTemplateId `
+                | Where-Object { $_.RoleTemplateId -eq "f2ef992c-3afb-46b9-b7cf-a126ee74c451" }
+            $globalReaders = $globalReaderRole.Members.Id
+            if ($globalReaders -contains $spnId)
+            {
+                Write-Host "Role granted." -ForegroundColor Green
+            }
+            else
+            {
+                throw "Error when trying to grant Global Reader role"
+            }
         }
         else
         {
-            Write-Host "Role was already granted." -ForegroundColor Green            
+            Write-Host "Role was already granted before." -ForegroundColor Green            
         }        
     }
     catch
@@ -1039,6 +1107,7 @@ if ("Y", "y" -contains $continueInput) {
         Write-Host $Error[0] -ForegroundColor Yellow
         Write-Host "Could not grant role. If you want Azure AD-based recommendations, please grant the Global Reader role manually to the $automationAccountName managed identity or, for previous versions of AOE, to the Run As Account principal." -ForegroundColor Red
     }
+    #endregion
 
     Write-Host "Deployment completed!" -ForegroundColor Green
 }
